requirements.txt
fastapi
uvicorn[standard]
twilio
gspread
google-auth
python-dotenv
